feat: #2 create flux of transfer cash between users

+++ database/migrations/2022_05_04_221638_create_wallets_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained('users');
-            $table->float('cash')->default(0);
+            $table->decimal('cash', 12, 2)->default(0);
             $table->timestamps();
         });
     }
